backend/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# Code analysis
radon>=5.1.0
pylint>=2.17.0
bandit>=1.7.5
cognitive-complexity>=1.3.0

# AI model integration
ollama>=0.1.0
huggingface-hub>=0.16.0
transformers>=4.30.0
torch>=2.0.0
aiohttp>=3.8.0

# Execution sandbox
docker>=6.1.0
psutil>=5.9.0

# Web dashboard
flask>=2.3.0
flask-cors>=4.0.0
plotly>=5.14.0
matplotlib>=3.7.0

# Reporting
reportlab>=4.0.0  # For PDF generation
jinja2>=3.1.0     # For HTML templates

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Utilities
gitpython>=3.1.0
python-dateutil>=2.8.0
tenacity>=8.2.0